requirements.txt
APScheduler==3.0.5
numpy>=1.10.4
oauthlib==1.0.3
Pillow==3.1.0
requests==2.9.1
requests-oauthlib==0.6.0
six==1.10.0
tweepy==3.5.0
wheel==0.24.0
wordcloud==1.2
